# src/Back-end/RiddleMeThis/RiddleMeThis/Dto/Usuario.cs
using System.ComponentModel.DataAnnotations;
namespace RiddleMeThis.Dto
{
    public class UsuarioDto
    {
        [Required]
        public int id { get; set; }

        [Required]
        [MaxLength(100)]
        public string? nome { get; set; }

        [Required]
        public int quizzes_feitos { get; set; }

    }
}