# src/Back-end/RiddleMeThis/RiddleMeThis/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Servidor API",
    "version": "v1"
  },
  "paths": {
    "/api/ControllerQuiz/ObterQuizAleatorio": {
      "get": {
        "tags": [
          "ControllerQuiz"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "dificuldade",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ControllerQuiz/AtualizarQuizzesFeitos": {
      "post": {
        "tags": [
          "ControllerQuiz"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/ControllerQuiz/ObterContagemQuizzesFeitos": {
      "get": {
        "tags": [
          "ControllerQuiz"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "usuarioId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Servidor/Cadastrar": {
      "post": {
        "tags": [
          "Servidor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Usuario"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Servidor/Login": {
      "post": {
        "tags": [
          "Servidor"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UsuarioDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Servidor/ObterUsuario": {
      "get": {
        "tags": [
          "Servidor"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "email",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "Usuario": {
      "required": [
        "nome"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        }
      }
    },
    "UsuarioDto": {
      "required": [
        "id",
        "nome",
        "quizzes_feitos"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "nome": {
          "maxLength": 100,
          "minLength": 1,
          "type": "string"
        },
        "quizzes_feitos": {
          "format": "int32",
          "type": "integer"
        }
      }
    }
  }
}